Fix admin setup
