+++ collectivo_app/collectivo.yml
@@ -8,6 +8,9 @@ secret_key: ${SECRET_KEY}
 allowed_hosts: ${ALLOWED_HOSTS}
 allowed_origins: ${ALLOWED_ORIGINS}
 
+admin_user: ${ADMIN_USER}
+admin_pass: ${ADMIN_PASS}
+
 development: true
 example_data: true
 api_docs: true
